+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- feat(shortcuts): add new menu actions and their default accelerators. (`809c61d`)
+
 ## 1.9.1 - 2026-06-30
 
 - feat(build): enhance electron-builder configuration and Vite setup. (`be1fef7`)